+++ fissure/Dashboard/Slots/TSITabSlots/__init__.py
@@ -2,5 +2,6 @@ from .detector import *
 from .conditioner import *
 from .feature_extractor import *
 from .sois import *
+from .survey import *
 # from .classifier import *
 from .legacy import *